Move the health handler into support
