+++ support/render/healthjson/response.go
@@ -1,0 +1,5 @@
+package healthjson
+
+type Response struct {
+	Status Status `json:"status"`
+}
