+++ support/render/healthjson/status.go
@@ -1,0 +1,7 @@
+package healthjson
+
+type Status string
+
+const (
+	StatusPass Status = "pass"
+)
